major refactor for simpler gallery data
