--- src/routes/api/images/generate/types.ts
@@ -1,7 +1,0 @@
-export interface GenerateImageRequest {
-  plantId: string;
-  description: string;
-  instructions?: string;
-  model?: string;
-  backgroundSecret?: string;
-}
