+++ src/routes/api/plants/[id]/candidateImage/+server.ts
@@ -5,10 +5,15 @@ import { eq } from "drizzle-orm";
 import { v4 as uuidv4 } from "uuid";
 import { uploadToS3 } from "$lib/server/images";
 import { URL_PREFIX } from "$lib/constants";
+import type { GeneratedImage } from "$lib/types";
 
-interface CandidateImageBody {
-  url?: string;
-  errorMessage?: string;
+/**
+ * Should be identical to the version in
+ * `netlify/functions/img-gen-background.mts`
+ */
+interface GenerateImageResultBody {
+  url?: string | null;
+  errorMessage?: string | null;
 }
 
 export const GET: RequestHandler = async ({ params }) => {
@@ -21,8 +26,8 @@ export const GET: RequestHandler = async ({ params }) => {
     if (candidateImage) {
       return json(candidateImage, { status: 200 });
     } else {
-      console.log("Candidate image not found; this is OK");
-      return json({}, { status: 202 });
+      // Not found. This could be because the image has not yet been created.
+      return error(404);
     }
   } else {
     return error(400, "plantId param required");
@@ -32,8 +37,9 @@ export const GET: RequestHandler = async ({ params }) => {
 export const POST: RequestHandler = async ({ request, params }) => {
   console.log("candidate image POST request");
   const plantId = params["id"];
-  const { url, errorMessage } = (await request.json()) as CandidateImageBody;
-  console.log("got", { url, plantId });
+  const { url, errorMessage } =
+    (await request.json()) as GenerateImageResultBody;
+  console.log("got", { url, plantId, errorMessage });
   if (plantId) {
     console.log({ url, errorMessage, plantId });
 
@@ -59,9 +65,11 @@ export const POST: RequestHandler = async ({ request, params }) => {
       } else {
         return error(500, "Failed to insert new generated image URL");
       }
+    } else {
+      // The response is "OK" but we had no URL
+      console.error("No URL; errorMessage ==", errorMessage);
+      return json({ errorMessage, plantId }, { status: 200 });
     }
-    // The response is "OK" but we had no URL
-    return json({ url, errorMessage, plantId }, { status: 200 });
   } else {
     return error(400, "plantId param required");
   }
