+++ src/routes/app/gallery/+page.server.ts
@@ -2,14 +2,11 @@ import type { PageServerLoad } from "./$types";
 import { getUserGarden, getUserSeeds } from "$lib/server";
 import { fail, redirect, type Actions } from "@sveltejs/kit";
 import { lucia } from "$lib/server/auth";
-import type { GardenViewData } from "$lib/types";
 import { db } from "$lib/server/db";
 import { eq } from "drizzle-orm";
 import { users } from "$lib/server/schema";
 
-export const load: PageServerLoad = async ({
-  locals
-}): Promise<GardenViewData> => {
+export const load: PageServerLoad = async ({ locals }) => {
   const username = locals.user?.username;
   const userId = locals.user?.id;
   if (!username) {
@@ -17,28 +14,23 @@ export const load: PageServerLoad = async ({
     redirect(302, "/app");
   }
 
-  console.log("gallery ******** (re)load page data");
-  if (userId) {
-    const garden = await getUserGarden(userId);
-    const seedBank = await getUserSeeds(userId);
-    console.log("...ready to render");
-    const thisUser = await db.query.users.findFirst({
-      where: eq(users.id, userId),
-      with: { mySeedbank: true }
-    });
-    if (thisUser) {
-      const viewData: GardenViewData = {
-        user: thisUser,
-        seedbank: seedBank,
-        garden
-      };
-      return viewData;
-    } else {
-      throw Error("could not find user when querying");
-    }
-  } else {
-    throw Error("userId missing");
+  if (!userId) {
+    throw Error("no user ID");
   }
+
+  const userWithSeedbankPlants = await db.query.users.findFirst({
+    where: eq(users.id, userId),
+    with: {
+      myGarden: true,
+      mySeedbank: { with: { plantsInSeedbank: { with: { plant: true } } } }
+    }
+  });
+
+  if (!userWithSeedbankPlants) {
+    throw Error("failed to load user with seedbank+plants");
+  }
+
+  return { userWithSeedbankPlants };
 };
 
 export const actions: Actions = {
