+++ src/routes/app/gallery/+page.svelte
@@ -1,7 +1,6 @@
 <script lang="ts">
   import { goto } from "$app/navigation";
   import {
-    type GardenViewData,
     type InsertPlant,
     type PlantWithDate,
     type SelectPlant
@@ -21,63 +20,59 @@
     MAX_CANVASSES,
     PLUG_NAMES
   } from "$lib/constants";
-  import PlantWasAddedPopup from "./pollination/PlantWasAddedPopup.svelte";
+  import PlantWasAddedPopup from "./PlantWasAddedPopup.svelte";
 
-  function addTimeLeft(inputData: GardenViewData) {
-    return {
-      ...inputData,
-      seedbank: {
-        ...inputData.seedbank,
-        plantsInSeedbank: inputData.seedbank.plantsInSeedbank.map((p) => ({
-          ...p,
-          plant: {
-            ...p.plant
-          }
-        }))
-      },
-      garden: {
-        ...inputData.garden,
-        plantsInGarden: [
-          // ...inputData.garden.plantsInGarden
-          ...inputData.garden.plantsInGarden.map((p) => ({
-            ...p,
-            timeLeft: DateTime.fromJSDate(p.pollinationDate)
-              .plus(DURATION_TILL_FERTILE)
-              .diffNow()
-          }))
-        ]
-      }
-    };
-  }
+  export let data;
+  type GalleryViewData = typeof data;
 
-  export let data: GardenViewData;
+  // function addTimeLeft(inputData: GalleryViewData) {
+  //   return {
+  //     ...inputData.userWithSeedbankPlants,
+  //     seedbank: {
+  //       ...inputData,
+  //       plantsInSeedbank: inputData.seedbank.plantsInSeedbank.map((p) => ({
+  //         ...p,
+  //         plant: {
+  //           ...p.plant,
+  //           timeLeft: DateTime.fromJSDate(p.pollinationDate)
+  //             .plus(DURATION_TILL_FERTILE)
+  //             .diffNow()
+  //         }))
+  //         }
+  //       }))
+  //     },
+  //     garden: {
+  //       ...inputData.garden,
+  //       plantsInGarden: [
+  //         // ...inputData.garden.plantsInGarden
+  //         ...inputData.garden.plantsInGarden.map((p) => ({
+  //           ...p,
+  //       ]
+  //     }
+  //   };
+  // }
 
-  let dataWithTimes = addTimeLeft(data);
-  console.log(dataWithTimes);
+  // let dataWithTimes = addTimeLeft(data);
 
-  let selectedPlant: PlantWithDate | null = null;
-
-  function handleClick(id: string) {
-    goto(`gallery/pollination/` + id);
-  }
+  let selectedPlant: SelectPlant | null = null;
 
   let yourOriginalPlant: SelectPlant | null =
-    data.seedbank.plantsInSeedbank.find(
+    data.userWithSeedbankPlants.mySeedbank.plantsInSeedbank.find(
       (plant) => plant.plant.parent1 == null && plant.plant.parent2 == null
     )?.plant || null;
 
-  function updateTimeLeft() {
-    dataWithTimes = addTimeLeft(data);
-  }
+  // function updateTimeLeft() {
+  //   dataWithTimes = addTimeLeft(data);
+  // }
 
   let agent: TetherAgent | null;
 
-  let datesInterval: NodeJS.Timeout | null = null;
+  // let datesInterval: NodeJS.Timeout | null = null;
 
   let newPlantForPopup: InsertPlant | null = null;
 
   onMount(async () => {
-    datesInterval = setInterval(updateTimeLeft, 1000);
+    // datesInterval = setInterval(updateTimeLeft, 1000);
     // const databaseInterval = setInterval(fetchdata, 1000);
 
     agent = await TetherAgent.create("app", {
@@ -90,8 +85,8 @@
     eventsPlug.on("message", (payload) => {
       const m = decode(payload) as EventNewPollination;
       if (
-        m.payload.authorTop === data.user.id ||
-        m.payload.authorBottom === data.user.id
+        m.payload.authorTop === data.userWithSeedbankPlants.id ||
+        m.payload.authorBottom === data.userWithSeedbankPlants.id
       ) {
         console.log("This plant belongs to me!", m.payload);
         newPlantForPopup = m.payload;
@@ -107,9 +102,9 @@
   });
 
   onDestroy(() => {
-    if (datesInterval) {
-      clearInterval(datesInterval);
-    }
+    // if (datesInterval) {
+    //   clearInterval(datesInterval);
+    // }
     if (agent) {
       agent.disconnect();
     }
@@ -124,58 +119,35 @@
   <div
     class="text-roel_blue font-primer text-3xl text-center border-b-[3px] border-roel_blue rounded-lg px-4 py-2"
   >
-    {data.garden.name}
+    {data.userWithSeedbankPlants.myGarden.name}
   </div>
-  {#each dataWithTimes.garden.plantsInGarden as plant, index}
+  {#each data.userWithSeedbankPlants.mySeedbank.plantsInSeedbank as plant, index}
     <!-- svelte-ignore a11y-click-events-have-key-events -->
     <!-- svelte-ignore a11y-no-static-element-interactions -->
     <!-- <code>{plant.plant.timeLeft.toFormat("mm:ss")}</code> -->
-    {#if plant.parent1 == null || plant.timeLeft.milliseconds < 0}
-      <div
-        on:click={() => {
-          console.log("click!");
-          selectedPlant = plant;
-        }}
-        class="cursor-pointer mt-4"
+    <div
+      on:click={() => {
+        console.log("click!");
+        selectedPlant = plant.plant;
+      }}
+      class="cursor-pointer mt-4"
+    >
+      <PlantDisplay
+        disableAnimation={index > MAX_CANVASSES - 1}
+        imageUrl={plant.plant.imageUrl || ""}
+        applyFilters={false}
+      />
+    </div>
+    <div class="mt-4 text-center">
+      <a
+        href={`/app/gallery/scan/${plant.plant.id}`}
+        data-test="start-pollinating-button"
+        data-umami-event="Start Pollinating Button"
+        class="bg-roel_blue text-roel_green font-primer text-3xl px-4 py-[0.5rem] mb-5 border-2 w-11/12 max-w-xs border-roel_blue rounded-full active:bg-roel_blue active:text-roel_green"
       >
-        <PlantDisplay
-          disableAnimation={index > MAX_CANVASSES - 1}
-          imageUrl={plant.imageUrl || ""}
-          applyFilters={false}
-        />
-      </div>
-      <div class="mt-4 text-center">
-        <button
-          data-test="start-pollinating-button"
-          data-umami-event="Start Pollinating Button"
-          class="bg-roel_blue text-roel_green font-primer text-3xl px-4 py-[0.5rem] mb-5 border-2 w-11/12 max-w-xs border-roel_blue rounded-full active:bg-roel_blue active:text-roel_green"
-          on:click={() => handleClick(plant.id)}
-        >
-          Start Pollinating
-        </button>
-      </div>
-    {:else}
-      <div
-        on:click={() => {
-          console.log("click!");
-          selectedPlant = plant;
-        }}
-        class="cursor-pointer mt-4"
-      >
-        <PlantDisplay
-          disableAnimation={index > MAX_CANVASSES - 1}
-          imageUrl={plant.imageUrl || ""}
-          applyFilters={true}
-        />
-      </div>
-      <div class="mt-4 text-center">
-        <button
-          class="bg-roel_green text-roel_blue font-primer text-2xl px-4 py-2 w-11/12 max-w-xs border-roel_blue border-[3px] rounded-full text-opacity-100"
-        >
-          Fertile in {plant.timeLeft.toFormat("mm:ss")}
-        </button>
-      </div>
-    {/if}
+        Start Pollinating
+      </a>
+    </div>
   {/each}
   <br />
   <br />
@@ -188,8 +160,6 @@
       selectedPlant = null;
     }}
     isOriginalPlant={selectedPlant.id == yourOriginalPlant?.id}
-    isPollinatingPlant={DateTime.fromJSDate(
-      selectedPlant.pollinationDate
-    ).diffNow() > DURATION_TILL_FERTILE}
+    isPollinatingPlant={true}
   ></PopupInfo>
 {/if}
