+++ src/routes/app/gallery/PlantWasAddedPopup.svelte
