+++ src/routes/app/gallery/newplant/+page.server.ts
