+++ src/routes/app/gallery/newplant/+page.svelte
